Module libimagentrytag::ui

Functions

get_add_tags
Get the tags which should be added from the commandline

get_remove_tags
Get the tags which should be removed from the commandline

tag_add_arg

tag_argument
Generates a clap::Arg which can be integrated into the commandline-ui builder for building a "-t" or "--tags" argument which takes values for tagging actions (add, remove)

tag_argument_name

tag_remove_arg

tag_subcommand
Generates a clap::SubCommand to be integrated in the commandline-ui builder for building a "tags --add foo --remove bar" subcommand to do tagging action.

tag_subcommand_add_arg_name

tag_subcommand_name

tag_subcommand_names

tag_subcommand_remove_arg_name